refactor: split macros implementation into multiple files
